+++ src/Janus.Hosting/Bff/ApiStatus.cs
@@ -37,6 +37,7 @@ internal static class ApiStatus
         [ErrorCodes.StartupRelyingPartyId] = StatusCodes.Status500InternalServerError,
         [ErrorCodes.StartupLabelLimit] = StatusCodes.Status500InternalServerError,
         [ErrorCodes.StartupRedirectClient] = StatusCodes.Status500InternalServerError,
+        [ErrorCodes.StartupSchemaMismatch] = StatusCodes.Status500InternalServerError,
         [ErrorCodes.PolicyUnregistered] = StatusCodes.Status500InternalServerError,
         [ErrorCodes.DerivationSourcesMissing] = StatusCodes.Status500InternalServerError,
         [ErrorCodes.SystemFault] = StatusCodes.Status500InternalServerError,